Fix casting

+++ rng.h
@@ -284,11 +284,11 @@ static inline int xNextIntJBetween(Xoroshiro *xr, const int min, const int max)
 
 static inline Xoroshiro xAtPos(Xoroshiro *xr, int x, int y, int z)
 {
-    uint64_t l = x * 3129871 ^ z * 116129781L ^ y;
+    int64_t l = (int64_t)(x * 3129871) ^ (int64_t)z * 116129781L ^ (int64_t)y;
     l = l * l * 42317861L + l * 11L;
     l >>= 16;
 
-    return (Xoroshiro) {l ^ xr->lo, xr->hi};
+    return (Xoroshiro) {(uint64_t)l ^ xr->lo, xr->hi};
 }
 
 // expand as necessary
